+++ api/helpers/analytics/showtime.js
@@ -63,6 +63,13 @@ module.exports = {
     records.map(record => {
       DJs[record.ID] = {
         name: record.name,
+        lastTwoWeeks: {
+          offStarts: [],
+          offEnds: [],
+          absences: [],
+          cancellations: [],
+          missedIDs: []
+        },
         semester: {
           showtime: 0,
           listeners: 0,
@@ -177,6 +184,7 @@ module.exports = {
             }
             DJs[record.dj].overall.missedIDs += record.missedIDs
             DJs[0].overall.missedIDs += record.missedIDs
+            if (moment().subtract(2, 'weeks').isBefore(record.actualStart)) { DJs[record.dj].lastTwoWeeks.missedIDs.push(moment(record.actualStart).toISOString(true)) }
             if (moment(sails.config.custom.startOfSemester).isBefore(moment(record.createdAt))) {
               if (record.ignore !== 1) {
                 DJs[record.dj].semester.reputationScore -= (2 * record.missedIDs)
@@ -218,12 +226,14 @@ module.exports = {
                   DJs[record.dj].overall.offStart += 1
                   if (record.ignore === 0) { DJs[0].overall.reputationScore -= 1 }
                   DJs[0].overall.offStart += 1
+                  if (moment().subtract(2, 'weeks').isBefore(record.actualStart)) { DJs[record.dj].lastTwoWeeks.offStarts.push(moment(record.actualStart).toISOString(true)) }
                 }
                 if (Math.abs(moment(record.scheduledEnd).diff(moment(record.actualEnd), 'minutes')) >= 10 && record.ignore !== 2) {
                   if (record.ignore === 0) { DJs[record.dj].overall.reputationScore -= 1 }
                   DJs[record.dj].overall.offEnd += 1
                   if (record.ignore === 0) { DJs[0].overall.reputationScore -= 1 }
                   DJs[0].overall.offEnd += 1
+                  if (moment().subtract(2, 'weeks').isBefore(record.actualStart)) { DJs[record.dj].lastTwoWeeks.offEnds.push(moment(record.actualStart).toISOString(true)) }
                 }
               }
               if (moment(sails.config.custom.startOfSemester).isBefore(moment(record.createdAt))) {
@@ -285,12 +295,14 @@ module.exports = {
                     DJs[record.dj].semester.offStart += 1
                     if (record.ignore === 0) { DJs[0].semester.reputationScore -= 1 }
                     DJs[0].semester.offStart += 1
+                    if (moment().subtract(2, 'weeks').isBefore(record.actualStart)) { DJs[record.dj].lastTwoWeeks.offStarts.push(moment(record.actualStart).toISOString(true)) }
                   }
                   if (Math.abs(moment(record.scheduledEnd).diff(moment(record.actualEnd), 'minutes')) >= 10 && record.ignore !== 2) {
                     if (record.ignore === 0) { DJs[record.dj].semester.reputationScore -= 1 }
                     DJs[record.dj].semester.offEnd += 1
                     if (record.ignore === 0) { DJs[0].semester.reputationScore -= 1 }
                     DJs[0].semester.offEnd += 1
+                    if (moment().subtract(2, 'weeks').isBefore(record.actualStart)) { DJs[record.dj].lastTwoWeeks.offEnds.push(moment(record.actualStart).toISOString(true)) }
                   }
                 }
               }
@@ -302,6 +314,7 @@ module.exports = {
               DJs[0].overall.reputationScore -= 1
               DJs[record.dj].overall.cancellations += 1
               DJs[0].overall.cancellations += 1
+              if (moment().subtract(2, 'weeks').isBefore(record.actualStart)) { DJs[record.dj].lastTwoWeeks.cancellations.push(moment(record.actualStart).toISOString(true)) }
             } else {
               if (record.happened === 0 && record.ignore !== 2) {
                 if (record.ignore === 0) {
@@ -310,6 +323,7 @@ module.exports = {
                 }
                 DJs[record.dj].overall.absences += 1
                 DJs[0].overall.absences += 1
+                if (moment().subtract(2, 'weeks').isBefore(record.actualStart)) { DJs[record.dj].lastTwoWeeks.absences.push(moment(record.actualStart).toISOString(true)) }
               } else if (record.happened === -1 && record.ignore !== 2) {
                 if (record.ignore === 0) {
                   DJs[record.dj].overall.reputationScore -= 1